move platform info component to components folder and delete obsolete formally_pages directory
