+++ src/components/PlatformInfo/platformInfo.tsx
