+++ src/components/PlatformInfo/types.tsx
